Revert "Merge commit 'ab61fb58' into msm-4.9 - PC232"
This reverts commit fd378b0ce2227beeffe33764901ecb11a4dfc8fd, reversing
changes made to a212cf530acf4a7d3b4b245755d69806e56df334.

+++ drivers/media/platform/msm/camera/cam_lrme/cam_lrme_context.c
@@ -22,7 +22,7 @@ static int __cam_lrme_ctx_acquire_dev_in_available(struct cam_context *ctx,
 	struct cam_acquire_dev_cmd *cmd)
 {
 	int rc = 0;
-	uintptr_t ctxt_to_hw_map = (uintptr_t)ctx->ctxt_to_hw_map;
+	uint64_t ctxt_to_hw_map = (uint64_t)ctx->ctxt_to_hw_map;
 	struct cam_lrme_context *lrme_ctx = ctx->ctx_priv;
 
 	CAM_DBG(CAM_LRME, "Enter ctx %d", ctx->ctx_id);
